garnishedvector->augvec
doesnt help

root has internal 1024 char limit on class names.  it is not clear how
this interacts with instantiated templates.  Shortening
GarnishedVector to AugVec appears to make the crash MORE persistent
rather than less.  Hopefully it is not something along the lines of
that rootcint dumps core if your class name is over 300 characters,
contains two or more templates, and has a checksum with is divisible by 11.

feh.


git-svn-id: http://code.icecube.wisc.edu/svn/dataclasses/trunk@696 16731396-06f5-0310-8873-f7f720988828