TriggerKey enum redo.  And fixing any code that had these enums used. There is a chance this might not be fully backward compatible, but the enum values I removed are ones that are likely not in wide use.
Note for the TWRreader.  Need to find a way to consistently
handle the M24 trigger pulse from the mudaq. As best I know
this is a flag a special channel has more timing information
and to look there.

Need a common sync timing solution that works for TWR/Mu daq
and IceCube/TWR daq.


git-svn-id: http://code.icecube.wisc.edu/svn/projects/dataclasses/trunk@20673 16731396-06f5-0310-8873-f7f720988828