***************************************************************** ***************************************************************** ************                DANGER                *************** ***************************************************************** *****************************************************************
Commit after the fact to released software.

Remove 'weight' member which does not belong here.

Sim production (already agreed with paolo and juan carlos) will
restart.  The files that *have* been written with this software will
still be readable, and you can just run them through a simple
reader-writer tray to sanitize them.  The problems would come when
data that was written by this software was not deleted, and then in
the future the I3Particle serialization is bumped again to version 4,
but the additional fields in version 4 are not 'weight'.  Then you'll
get crashes.

Short story:  run the data that was written with this release through
a simple reader-writer tray to sanitize, or delete them.





git-svn-id: http://code.icecube.wisc.edu/svn/projects/dataclasses/releases/V02-02-01@48739 16731396-06f5-0310-8873-f7f720988828