Added more pythonic bindings for the std::map-derived classes.
The wrappers for std::map-like classes now support almost all of the
methods of Python dictionaries. Whereas with the standard wrappers one
had to extract the keys from a map by hand with a list comprehension:

keys = [pair.key() for pair in i3map]

one can now do this in a single statement:

keys = i3map.keys()

If all you want is a list of the keys, then the latter can be about
8x faster. Iteration over the map still returns wrapped std::pairs for
backwards compatibility, but these can be converted to tuples or unpacked
implicitly. Whereas before one had to unpack arguments inside loops:

for pair in frame['I3Geometry'].omgeo:
	om_key = pair.key()
	om_geometry = pair.data()
	# do stuff with geometry

one can now save a bit of typing:

for om_key,om_geometry in frame['I3Geometry'].omgeo:
	# do stuff right away!

Note: the tuple-like interface for the std::pairs is a bit hacky and
slow. At the moment it is significantly faster to use the iterator form:

for om_key,om_geometry in frame['I3Geometry'].omgeo.iteritems():
	# do stuff slightly faster


git-svn-id: http://code.icecube.wisc.edu/svn/projects/dataclasses/trunk@56079 16731396-06f5-0310-8873-f7f720988828