Change definition of I3Waveform::Status to be more useful and intelligible to humans.
I3Waveform::Status has been replaced with an anonymous enum. VIRGINAL has remained, the old SHADY is now COMBINED, ADULTERATED is now SATURATED, and a new value, UNDERSHOT, was added. I3Waveform::GetStatus() now combines these with a bitwise or, so an FADC waveform that saturates at the beginning and undershoots its bias voltage in the middle will be marked SATURATED | UNDERSHOT.

git-svn-id: http://code.icecube.wisc.edu/svn/projects/dataclasses/trunk@71849 16731396-06f5-0310-8873-f7f720988828