Add time-shifting to I3RecoPulseSeriesMapMask, bumping the version to 1.
This makes it possible for I3TriggerSplitter to shift pulses in sub-events
so that their times are defined with respect to the beginning of the new
readout window. Client code dealing with split events can then continue
to expect all pulses to fall within the readout window whether trigger
splitting has been done or not.


git-svn-id: http://code.icecube.wisc.edu/svn/projects/dataclasses/trunk@90189 16731396-06f5-0310-8873-f7f720988828