Bring forward r90189 to RC:
This makes it possible for I3TriggerSplitter to shift pulses in sub-events
so that their times are defined with respect to the beginning of the new
readout window. Client code dealing with split events can then continue
to expect all pulses to fall within the readout window whether trigger
splitting has been done or not.

Also picked up some clang fixes (in the {} placement) as well.  


git-svn-id: http://code.icecube.wisc.edu/svn/projects/dataclasses/releases/V12-06-00@90219 16731396-06f5-0310-8873-f7f720988828